requirements.txt
gunicorn
requests
flask
dadjokes
mysql-connector
PyMySQL
Flask-MySQL